// packages/ui/src/components/ColorChips/ColorChips.types.ts
// Generated with util/create-component.js
export interface ColorChipsProps {
  colors: string;
  chipWidth: string;
  chipHeight: string;
  direction?: "row" | "column";
  shape?: "circle" | "square";
  size?: number;
}

// color chip (product color variance)
// silver
export const COLOR_CHIP_SILVER = "#AAAAAA";
export const COLOR_CHIP_SILVER_LIGHT = "#E5E5E5";
export const COLOR_CHIP_SILVER_DARK = "#6A6A6A";

// beige
export const COLOR_CHIP_BEIGE = "#C7B3A1";
export const COLOR_CHIP_BEIGE_LIGHT = "#E8DBCF";
export const COLOR_CHIP_BEIGE_DARK = "#8C7967";

// blue
export const COLOR_CHIP_BLUE = "#0085FF";
export const COLOR_CHIP_BLUE_LIGHT = "#B9E6FF";
export const COLOR_CHIP_BLUE_DARK = "#002ECF";

// navy
export const COLOR_CHIP_NAVY = "#001E87";
export const COLOR_CHIP_NAVY_LIGHT = "#6678B8";
export const COLOR_CHIP_NAVY_DARK = "#001253";

// green
export const COLOR_CHIP_GREEN = "#00C566";
export const COLOR_CHIP_GREEN_LIGHT = "#49FF7C";
export const COLOR_CHIP_GREEN_DARK = "#00793F";

// red
export const COLOR_CHIP_RED = "#FF4037";
export const COLOR_CHIP_RED_LIGHT = "#FF7F6D";
export const COLOR_CHIP_RED_DARK = "#951C16";

// yellow
export const COLOR_CHIP_YELLOW = "#FFE358";
export const COLOR_CHIP_YELLOW_LIGHT = "#FFF3B5";
export const COLOR_CHIP_YELLOW_DARK = "#E9C200";

// mint
export const COLOR_CHIP_MINT = "#32FFCE";
export const COLOR_CHIP_MINT_LIGHT = "#C0FFF0";
export const COLOR_CHIP_MINT_DARK = "#00CEB5";

// pink
export const COLOR_CHIP_PINK = "#FF79B1";
export const COLOR_CHIP_PINK_LIGHT = "#FFD7E7";
export const COLOR_CHIP_PINK_DARK = "#E1166A";

// purple
export const COLOR_CHIP_PURPLE = "#A653E8";
export const COLOR_CHIP_PURPLE_LIGHT = "#DFBEF9";
export const COLOR_CHIP_PURPLE_DARK = "#5E0D9D";

// brown
export const COLOR_CHIP_BROWN = "#955F0E";
export const COLOR_CHIP_BROWN_LIGHT = "#C89A57";
export const COLOR_CHIP_BROWN_DARK = "#623812";

// orange
export const COLOR_CHIP_ORANGE = "#FF782D";
export const COLOR_CHIP_ORANGE_LIGHT = "#FF9F2F";
export const COLOR_CHIP_ORANGE_DARK = "#D24C00";

// gold
export const COLOR_CHIP_GOLD = "#DFA614";
export const COLOR_CHIP_GLOD_LIGHT = "#EFD081";
export const COLOR_CHIP_GOLD_DARK = "#C38100";

// pattern
export const COLOR_CHIP_PATTERN = ["#000000", "#83D2FF", "#83D2FF", "#000000"];

export const COLOR_CHIPS = {
  화이트: "#FFFFFF",
  블랙: "#000000",

  실버: COLOR_CHIP_SILVER,
  라이트실버: COLOR_CHIP_SILVER_LIGHT,
  다크실버: COLOR_CHIP_SILVER_DARK,

  베이지: COLOR_CHIP_BEIGE,
  라이트베이지: COLOR_CHIP_BEIGE_LIGHT,
  다크베이지: COLOR_CHIP_BEIGE_DARK,

  블루: COLOR_CHIP_BLUE,
  라이트블루: COLOR_CHIP_BLUE_LIGHT,
  다크블루: COLOR_CHIP_BLUE_DARK,

  네이비: COLOR_CHIP_NAVY,
  라이트네이비: COLOR_CHIP_NAVY_LIGHT,
  다크네이비: COLOR_CHIP_NAVY_DARK,

  그린: COLOR_CHIP_GREEN,
  라이트그린: COLOR_CHIP_GREEN_LIGHT,
  다크그린: COLOR_CHIP_GREEN_DARK,

  레드: COLOR_CHIP_RED,
  라이트레드: COLOR_CHIP_RED_LIGHT,
  다크레드: COLOR_CHIP_RED_DARK,

  옐로우: COLOR_CHIP_YELLOW,
  라이트옐로우: COLOR_CHIP_YELLOW_LIGHT,
  다크옐로우: COLOR_CHIP_YELLOW_DARK,

  민트: COLOR_CHIP_MINT,
  라이트민트: COLOR_CHIP_MINT_LIGHT,
  다크민트: COLOR_CHIP_MINT_DARK,

  핑크: COLOR_CHIP_PINK,
  라이트핑크: COLOR_CHIP_PINK_LIGHT,
  다크핑크: COLOR_CHIP_PINK_DARK,

  퍼플: COLOR_CHIP_PURPLE,
  라이트퍼플: COLOR_CHIP_PURPLE_LIGHT,
  다크퍼플: COLOR_CHIP_PURPLE_DARK,

  브라운: COLOR_CHIP_BROWN,
  라이트브라운: COLOR_CHIP_BROWN_LIGHT,
  다크브라운: COLOR_CHIP_BROWN_DARK,

  오렌지: COLOR_CHIP_ORANGE,
  라이트오렌지: COLOR_CHIP_ORANGE_LIGHT,
  다크오렌지: COLOR_CHIP_ORANGE_DARK,

  골드: COLOR_CHIP_GOLD,
  라이트골드: COLOR_CHIP_GLOD_LIGHT,
  다크골드: COLOR_CHIP_GOLD_DARK,
};